API

Documentation for LPSjl.jl

LPSjl.DAG — Type

DAG

An Acyclic DiGraph

The edges and nodes are stored in LightGraphs.DiGraph but checks are are made to prevent loops.

Constructors

    DAG()
    DAG(digraph::DiGraph)
    DAG(nodes::Array{String})
    DAG(edges::Array{Tuple{String}})

Keyword arguments

digraph: an existing LightGraphs.DiGraph

nodes : array of nodes of the graph

edges : array of edges of the graph

LPSjl.is_dag — Method

is_dag(digraph::DiGraph)

Verify if a DiGraph is a DAG, throwing an error if it isn't

LPSjl.add_node! — Function

add_node!(dag::DAG, node::String)

Add a node to the DAG

LPSjl.add_edge! — Function

add_edge!(dag::DAG, node₁::String, node₂::String)

Add an edge between two nodes to the DAG, creating the non-existent nodes